Extract connecting to agent from MCP
This will be used in multiple tools.
